Move actionStatus property after ActionStatusType definition
Due to a quirk in the current processing approach, the ActionStatusType
definition will not be rendered correctly as an enumeration because
the actionStatus property that refers to it is defined earlier in schema.rdfa.
We can work on a more robust processing approach, but in the short term
the easiest fix is to simply move the definition to later in schema.rdfa.

Signed-off-by: Dan Scott <dan@coffeecode.net>