Consistency of encoding / associatedMedia
Use the same definition for encoding and associatedMedia,
as they are synonyms for one another. Also make the definition
agree grammatically with singular vs plural.

Also refer to "CreativeWork" in the description rather than
"creative work", as the former seems to be the predominant form used
for properties on CreativeWork.

Signed-off-by: Dan Scott <dan@coffeecode.net>