Refined flightNumber and Flight descriptions.
1.
Backed out suggested addition of "Use 'provider' to represent carrier" to Flight description.

2.
2.
Updated the description of 'flightNumber' to
"The unique identifier for a flight including the airline IATA code.
For example, if describing United flight 110, where the IATA code for United is 'UA',
the flightNumber is 'UA110'."

old: The unique identifier for a flight, not including the airline IATA code. For example, if describing United flight 110, the flightNumber is '110'. The IATA code can be set on the Airline.
new: The unique identifier for a flight including the airline IATA code. For example, if describing United flight 110, where the IATA code for United is 'UA', the flightNumber is 'UA110'.

This is a change in recommended use, see
https://www.w3.org/wiki/WebSchemas/SellerVendorCleanup
https://github.com/rvguha/schemaorg/pull/89
https://github.com/rvguha/schemaorg/pull/91 for background.