Moved terms around PhysicalActivity and ExercisePlan into health-lifesci. See #492 Also #1117 for general best practice documentation which is much needed!
This fixes and implementation issue discovered during v3.0 release candidate review.

We had terms in the leaves of the type hierarchy that were in core, but whose
supertypes were only defined in an extension. This does not work in our current
framework.

Details:
[core]Thing > [health-lifesci]MedicalEntity >[health-lifesci] MedicalTherapy >[health-lifesci] LifestyleModification > [core]PhysicalActivity >[core] ExercisePlan

Resolution (confirmed with @twamarc) was to move the subtrees across
into the health-lifesci extension. This makes sense also because the tone
of the documentation of these terms is in a professional/healthcare voice,
and the terms originated in the same 2013 collaboration (see docs/meddocs.html).

The ExercisePlan type also related to the ExerciseAction type in core. We
have moved the linking property here.

Finally, annotations on 'category' from core extend it to work with two
types from this extension.