Optometic -> Optometric. Per #492 unless there's a quick consensus on a better mapping.
As this is relatively obscure vocabulary I have here decided not to
use supersededBy and retain the old typo version. If it turns out the
typo had significant use we can restore the superseded version for the
records, but my sense is we just make the project look untidy for
minimal gain.