Fixed reversions to netWorth definition.
We choose not to mention organizations at this stage. FIBO and other
considerations could mean we revisit this, but this commit fixes an
accidental reversion that prematurely reintroduced idea that netWorth
applied to Organization. If/when we do that, we need to change domainIncludes also.

The 3.0 release had this reverted. See #1203

See also http://schema.org/docs/releases.html#g585 #585