We don't have length/thickness under Product
We need to clarify that depth is also known as thickness.
In expanded form of the depth description, we could even say 'the depth or thickness or length of the item'.  But this begs the question of how not to confuse thickness with height

In general terms, its left/right, front/backtop/bottom.

In the States, we typically express those same general term of a Products' dimensions as length/width/height  .. but for Product we currently have depth/width/height.  I think for developers, that causes more harm than good in trying to align with those general terms.