Responsiveness: specify padding in ems not px
If we work with a relative unit like ems then the design can be more
responsive. In this case, 40px was a huge amount of padding on smaller
screens; use 0.5em padding and increase that on larger displays.

Also, remove the CSS declarations that were made redundant by subsequent
declarations.

Signed-off-by: Dan Scott <dan@coffeecode.net>