Responsive design commit series (#1878)
* Responsiveness: remove global font-size: 75%

Use the default font size and allow users to adjust up or down if necessary.
75% can be quite hard to see on modern high-resolution screens, for example.

Signed-off-by: Dan Scott <dan@coffeecode.net>

* Responsiveness: set min-width and max-width only for large media

If the width of the display media is greater than 960px, then we can set the
min-width and max-width accordingly. Otherwise, let's leave that unspecified
and let the content respond to the the device.

Signed-off-by: Dan Scott <dan@coffeecode.net>

* Responsiveness: specify padding in ems not px

If we work with a relative unit like ems then the design can be more
responsive. In this case, 40px was a huge amount of padding on smaller
screens; use 0.5em padding and increase that on larger displays.

Also, remove the CSS declarations that were made redundant by subsequent
declarations.

Signed-off-by: Dan Scott <dan@coffeecode.net>

* Responsiveness: property table display

Allow table elements to display as inline-block elements on smaller screens.

Definitely not perfect, but allows the content to be viewed without pinching
and zooming or scrolling back and forth.

Signed-off-by: Dan Scott <dan@coffeecode.net>

* Responsiveness: less whitespace at page start

2em in an h1 is a lot of white space; cut that in half and it works reasonably
well on small and big screens alike.

Signed-off-by: Dan Scott <dan@coffeecode.net>

* Responsiveness: add a meta viewport

Also use generally accepted HTML5 html and meta elements, instead of XHTML, to
simplify and avoid potential quirks.

Signed-off-by: Dan Scott <dan@coffeecode.net>