Add a `snapshot_schema` file which snapshots the full schema in it's current state.
The snapshot_ttl function writes the current state of the schema into
some output dir, defaulting to `software/tests/snapshot/schemaorg-all-https.ttl`. This file can then be
used for a simple snapshot/diff test to avoid accidental changes to the
schema. When changing the schema the test fails and the 'golden' file needs to be
updated, the updated snapshot is then included into any changes.

MARKDOWN=true