Add schema snapshotting util and a snapshot test (#4594)
Adding:

1. minimal snapshotting utilities which exports the all-https turtle version of the current schema.
2. 'golden' file of the current schema
3. snapshot test which checks that the current snapshot is up to date with the current schema

This ensures that changes to the complete schema can be observed easily, including changes/additions to the schema itself as well as any code building the schema.

The test could be a fail fast test to run in CI. We can add an github action to run it separately in a follow up. The test runs with the rest of the tests already, but this is quite slow.